requirements.txt
numpy<2
pandas
scikit-learn
scikit-surprise
joblib
fastapi
uvicorn
scipy
sqlalchemy
aiosqlite
asyncpg
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
sentence-transformers
faiss-cpu
ollama 
tqd
aiofiles